Replace images in Drawing/Presentation documents
Can now replace an image in a Drawing/Presentation document, and save the
document in the original format (i.e. .ppt).

This shows that the outline information is lost for some reason.

Next step is to just open a .ppt document and save it to a new .ppt document
and determine whether that causes the same outline information loss.