Citation basics are complete
This includes basic functionality for citations for both Impress
and Writer documents.  Next step is to look into doing the
hyperlink method we discussed.  (However, Piet has indicated that
some other visual method (inserting a footnote or some other method
of visually matching the image to the citation information) is
required.

Perhaps the footnote can be the link rather than the image?