Working front boiler plate code
This works as far as copying the front boilerplate stuff into
the document.  Need to work on fixing the background, and moving
the original first slide after the copied stuff.

Wanted to commit this before cleaning up and possibly breaking
something!

The "magic" seems to have been adding the dispatch to get the
source file into DrawingMode...  :-/